test(client): move p2p tests into separate module
