+++ client/src/tests.rs
@@ -7,4 +7,6 @@ mod fresh;
 mod actor_tests;
 mod basic_tests;
 mod interface_tests;
+#[cfg(feature = "p2p")]
+mod p2p_tests;
 mod procedures_tests;
