+++ client/src/tests/interface_tests.rs
@@ -4,15 +4,6 @@
 use crate::{ActorError, Location, RecordHint, Stronghold};
 use stronghold_utils::random::bytestring;
 
-#[cfg(feature = "p2p")]
-use crate::{
-    p2p::{identity::Keypair, NetworkConfig, PeerId, Permissions, SwarmInfo},
-    procedures::{Slip10Derive, Slip10DeriveInput, Slip10Generate},
-    tests::fresh,
-};
-#[cfg(feature = "p2p")]
-use tokio::sync::{mpsc, oneshot};
-
 #[actix::test]
 async fn test_stronghold() {
     let vault_path = b"path".to_vec();
@@ -395,307 +386,4 @@ async fn test_stronghold_generics() {
         .await
         .unwrap_or_else(|e| panic!("Actor error: {}", e))
         .unwrap_or_else(|e| panic!("Write snapshot error: {}", e));
-}
-
-#[cfg(feature = "p2p")]
-#[actix::test]
-async fn test_stronghold_p2p() {
-    let system = actix::System::current();
-    let arbiter = system.arbiter();
-
-    let (addr_tx, addr_rx) = oneshot::channel();
-
-    // Channel for signaling that local/ remote is ready i.g. performed a necessary write, before the other ran try
-    // read.
-    let (remote_ready_tx, mut remote_ready_rx) = mpsc::channel(1);
-    let (local_ready_tx, mut local_ready_rx) = mpsc::channel(1);
-
-    let key1 = bytestring(1024);
-    let data1 = b"some data".to_vec();
-    let key1_clone = key1.clone();
-    let data1_clone = data1.clone();
-
-    let key2 = bytestring(1024);
-    let data2 = b"some second data".to_vec();
-    let key2_clone = key2.clone();
-    let data2_clone = data2.clone();
-
-    let seed1 = fresh::location();
-    let seed1_clone = seed1.clone();
-
-    let (res_tx, mut res_rx) = mpsc::channel(1);
-    let res_tx_clone = res_tx.clone();
-
-    let remote_client = b"remote".to_vec();
-    let remote_client_clone = remote_client.clone();
-
-    let spawned_local = arbiter.spawn(async move {
-        let local_client = b"local".to_vec();
-        let mut local_stronghold = Stronghold::init_stronghold_system(local_client, vec![])
-            .await
-            .unwrap_or_else(|e| panic!("Could not create a stronghold instance: {}", e));
-        local_stronghold
-            .spawn_p2p(NetworkConfig::new(Permissions::allow_all()), None)
-            .await
-            .unwrap_or_else(|e| panic!("Could not spawn p2p: {}", e));
-
-        let (peer_id, addr) = addr_rx.await.unwrap();
-        local_stronghold
-            .add_peer(peer_id, Some(addr))
-            .await
-            .unwrap_or_else(|e| panic!("Actor error: {}", e))
-            .unwrap_or_else(|e| panic!("Could not establish connection to remote: {}", e));
-
-        remote_ready_rx.recv().await.unwrap();
-
-        // test writing at remote and reading it from local stronghold
-        let payload = local_stronghold
-            .read_from_remote_store(peer_id, remote_client_clone.clone(), key1)
-            .await
-            .unwrap_or_else(|e| panic!("Could not read from remote store: {}", e));
-        assert_eq!(payload.unwrap(), data1);
-
-        // test writing from local and reading it at remote
-        local_stronghold
-            .write_to_remote_store(peer_id, remote_client_clone.clone(), key2, data2, None)
-            .await
-            .unwrap_or_else(|e| panic!("Could not write to remote store: {}", e));
-        local_ready_tx.send(()).await.unwrap();
-
-        // test writing and reading from local
-        let key3 = bytestring(1024);
-        let original_data3 = b"some third data".to_vec();
-        local_stronghold
-            .write_to_remote_store(
-                peer_id,
-                remote_client_clone.clone(),
-                key3.clone(),
-                original_data3.clone(),
-                None,
-            )
-            .await
-            .unwrap_or_else(|e| panic!("Could not write to remote store: {}", e));
-
-        let payload = local_stronghold
-            .read_from_remote_store(peer_id, remote_client_clone.clone(), key3)
-            .await
-            .unwrap_or_else(|e| panic!("Could not read from remote store: {}", e));
-
-        assert_eq!(payload.unwrap(), original_data3);
-
-        remote_ready_rx.recv().await.unwrap();
-
-        let (_path, chain) = fresh::hd_path();
-
-        match local_stronghold
-            .remote_runtime_exec(
-                peer_id,
-                remote_client_clone,
-                Slip10Derive {
-                    output: fresh::location(),
-                    chain,
-                    hint: fresh::record_hint(),
-                    input: Slip10DeriveInput::Seed(seed1),
-                },
-            )
-            .await
-            .unwrap_or_else(|e| panic!("Could not execute remote procedure: {}", e))
-        {
-            Ok(_) => {}
-            Err(e) => panic!("unexpected error: {:?}", e),
-        };
-        res_tx.send(()).await.unwrap();
-    });
-    assert!(spawned_local);
-
-    let spawned_remote = arbiter.spawn(async move {
-        let mut remote_stronghold = Stronghold::init_stronghold_system(remote_client, vec![])
-            .await
-            .unwrap_or_else(|e| panic!("Could not create a stronghold instance: {}", e));
-        remote_stronghold
-            .spawn_p2p(NetworkConfig::new(Permissions::allow_all()), None)
-            .await
-            .unwrap_or_else(|e| panic!("Could not create a stronghold instance: {}", e));
-
-        let addr = remote_stronghold
-            .start_listening(None)
-            .await
-            .unwrap_or_else(|e| panic!("Actor error: {}", e))
-            .unwrap_or_else(|e| panic!("Could not start listening: {}", e));
-
-        let swarm_info = remote_stronghold
-            .get_swarm_info()
-            .await
-            .unwrap_or_else(|e| panic!("Could not get swarm info: {}", e));
-
-        assert!(swarm_info.listeners.into_iter().any(|l| l.addrs.contains(&addr)));
-        addr_tx.send((swarm_info.local_peer_id, addr)).unwrap();
-        // test writing at remote and reading it from local stronghold
-        remote_stronghold
-            .write_to_store(key1_clone, data1_clone, None)
-            .await
-            .unwrap_or_else(|e| panic!("Could not write to remote store: {}", e));
-
-        remote_ready_tx.send(()).await.unwrap();
-        local_ready_rx.recv().await.unwrap();
-
-        // test writing from local and reading it at remote
-        let payload = remote_stronghold
-            .read_from_store(key2_clone)
-            .await
-            .unwrap_or_else(|e| panic!("Could not read from remote store: {}", e));
-        assert_eq!(payload.unwrap(), data2_clone);
-
-        // test procedure execution
-        match remote_stronghold
-            .runtime_exec(Slip10Generate {
-                size_bytes: None,
-                output: seed1_clone,
-                hint: fresh::record_hint(),
-            })
-            .await
-            .unwrap_or_else(|e| panic!("Could not execute remote procedure: {}", e))
-        {
-            Ok(_) => {}
-            Err(e) => panic!("unexpected error: {:?}", e),
-        };
-
-        remote_ready_tx.send(()).await.unwrap();
-        res_tx_clone.send(()).await.unwrap();
-    });
-    assert!(spawned_remote);
-
-    // wait for both threads to return
-    res_rx.recv().await.unwrap();
-    res_rx.recv().await.unwrap();
-}
-
-#[cfg(feature = "p2p")]
-#[actix::test]
-async fn test_p2p_config() {
-    use crate::p2p::{OutboundFailure, P2pError, Permissions};
-
-    let remote_client = bytestring(4096);
-    // Start remote stronghold and start listening
-    let mut remote_sh = Stronghold::init_stronghold_system(remote_client.clone(), vec![])
-        .await
-        .unwrap();
-    match remote_sh
-        .spawn_p2p(
-            NetworkConfig::new(Permissions::allow_all()).with_mdns_enabled(false),
-            None,
-        )
-        .await
-    {
-        Ok(()) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-    let remote_addr = match remote_sh.start_listening(None).await.unwrap() {
-        Ok(a) => a,
-        Err(e) => panic!("Unexpected error {}", e),
-    };
-    let SwarmInfo {
-        local_peer_id: remote_id,
-        ..
-    } = remote_sh.get_swarm_info().await.unwrap();
-
-    // Start (local) stronghold.
-    let client_path = fresh::bytestring(4096);
-    let mut stronghold = Stronghold::init_stronghold_system(client_path.clone(), vec![])
-        .await
-        .unwrap();
-    // Generate a new Keypair and write it to the vault
-    let keypair = Keypair::generate_ed25519();
-    let peer_id = PeerId::from_public_key(&keypair.public());
-    let keys_location = fresh::location();
-    match stronghold
-        .write_p2p_keypair(keypair, keys_location.clone(), fresh::record_hint())
-        .await
-        .unwrap()
-    {
-        Ok(()) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-    // Spawn p2p that uses the new keypair
-    match stronghold
-        .spawn_p2p(
-            NetworkConfig::new(Permissions::allow_all()).with_mdns_enabled(false),
-            Some(keys_location.clone()),
-        )
-        .await
-    {
-        Ok(()) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-
-    // Start listening
-    let addr = match stronghold.start_listening(None).await.unwrap() {
-        Ok(addr) => addr,
-        Err(e) => panic!("Unexpected error {}", e),
-    };
-    let SwarmInfo {
-        local_peer_id,
-        listeners,
-        ..
-    } = stronghold.get_swarm_info().await.unwrap();
-    assert_eq!(local_peer_id, peer_id);
-    assert!(listeners.first().unwrap().addrs.contains(&addr));
-    // Set a firewall rule
-    stronghold
-        .set_firewall_permissions(Permissions::default(), vec![remote_id], false)
-        .await
-        .unwrap();
-    // Add the remote's address info
-    match stronghold.add_peer(remote_id, Some(remote_addr.clone())).await.unwrap() {
-        Ok(_) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-    // Test that the firewall rule is effective
-    let res = remote_sh
-        .read_from_remote_store(peer_id, remote_client.clone(), bytestring(10))
-        .await;
-    match res {
-        Ok(_) => panic!("Request should be rejected."),
-        Err(P2pError::Local(e)) => panic!("Unexpected error {}", e),
-        Err(P2pError::SendRequest(OutboundFailure::NotPermitted))
-        | Err(P2pError::SendRequest(OutboundFailure::DialFailure))
-        | Err(P2pError::SendRequest(OutboundFailure::Shutdown))
-        | Err(P2pError::SendRequest(OutboundFailure::Timeout)) => panic!("Unexpected error {:?}", res),
-        Err(_) => {}
-    }
-
-    // Stop p2p and store the config in the stronghold store.
-    // This should persist firewall configuration and the collected address info about the remote.
-    let store = bytestring(1024);
-    match stronghold.stop_p2p(Some(store.clone())).await.unwrap() {
-        Ok(()) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-
-    // Spawn p2p again and load the config. Use the same keypair to keep the same peer-id.
-    match stronghold.spawn_p2p_load_config(store, Some(keys_location)).await {
-        Ok(()) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-    let SwarmInfo { local_peer_id, .. } = stronghold.get_swarm_info().await.unwrap();
-
-    // Test if the local peer still has the remote's address info.
-    assert_eq!(local_peer_id, peer_id);
-    match stronghold.add_peer(remote_id, None).await.unwrap() {
-        Ok(_) => {}
-        Err(e) => panic!("Unexpected error {}", e),
-    }
-    // Test that the firewall rule is still effective
-    let res = remote_sh
-        .read_from_remote_store(peer_id, remote_client, bytestring(10))
-        .await;
-    match res {
-        Ok(_) => panic!("Request should be rejected."),
-        Err(P2pError::Local(e)) => panic!("Unexpected error {}", e),
-        Err(P2pError::SendRequest(OutboundFailure::NotPermitted))
-        | Err(P2pError::SendRequest(OutboundFailure::DialFailure))
-        | Err(P2pError::SendRequest(OutboundFailure::Shutdown))
-        | Err(P2pError::SendRequest(OutboundFailure::Timeout)) => panic!("Unexpected error {:?}", res),
-        Err(_) => {}
-    }
 }
